==> 1_Setup_Dataset_Directories.ipynb <==
# Mount Google Drive

from google.colab import drive
drive.mount('/content/drive')

# Specify All The Directories

# Train set path
glioma_train_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Training/glioma_tumor'
meningioma_train_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Training/meningioma_tumor'
notumor_train_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Training/no_tumor'
pituitary_train_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Training/pituitary_tumor'

# Test set path
glioma_test_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Testing/glioma_tumor'
meningioma_test_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Testing/meningioma_tumor'
notumor_test_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Testing/no_tumor'
pituitary_test_path = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Testing/pituitary_tumor'

# Rename Images for Train Set

# Get a list of all files in the directory
files = os.listdir(glioma_train_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(glioma_train_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"glioma_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(glioma_train_path, filename), os.path.join(glioma_train_path, new_filename))

print("Train Set/Glioma_All files have been renamed.")

# Get a list of all files in the directory
files = os.listdir(meningioma_train_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(meningioma_train_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"meningioma_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(meningioma_train_path, filename), os.path.join(meningioma_train_path, new_filename))

print("Train Set/Meningioma_All files have been renamed.")

# Get a list of all files in the directory
files = os.listdir(notumor_train_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(notumor_train_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"notumor_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(notumor_train_path, filename), os.path.join(notumor_train_path, new_filename))

print("Train Set/NoTumor_All files have been renamed.")

# Get a list of all files in the directory
files = os.listdir(pituitary_train_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(pituitary_train_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"pituitary_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(pituitary_train_path, filename), os.path.join(pituitary_train_path, new_filename))

print("Train Set/Pituitary_All files have been renamed.")

# Rename Images for Test Set

# Get a list of all files in the directory
files = os.listdir(glioma_test_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(glioma_test_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"glioma_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(glioma_test_path, filename), os.path.join(glioma_test_path, new_filename))

print("Test Set/Glioma_All files have been renamed.")

# Get a list of all files in the directory
files = os.listdir(meningioma_test_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(meningioma_test_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"meningioma_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(meningioma_test_path, filename), os.path.join(meningioma_test_path, new_filename))

print("Test Set/Meningioma_All files have been renamed.")

# Get a list of all files in the directory
files = os.listdir(notumor_test_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(notumor_test_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"notumor_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(notumor_test_path, filename), os.path.join(notumor_test_path, new_filename))

print("Test Set/NoTumor_All files have been renamed.")

# Get a list of all files in the directory
files = os.listdir(pituitary_test_path)

# Iterate through each file and rename it
for filename in files:
    # Check if the file is a regular file (not a directory)
    if os.path.isfile(os.path.join(pituitary_test_path, filename)):
        # Split the file name and extension
        name, extension = os.path.splitext(filename)
        new_filename = f"pituitary_{name}{extension}"
        # Rename the file
        os.rename(os.path.join(pituitary_test_path, filename), os.path.join(pituitary_test_path, new_filename))

print("Test Set/Pituitary_All files have been renamed.")

# Define Train Directory

import os
import shutil

# Define the train directory
train_dir = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Train'

# Ensure the existence of the train and test directories
os.makedirs(train_dir, exist_ok=True)

# Move the remaining files to the train directory
for file in os.listdir(glioma_train_path):
    src = os.path.join(glioma_train_path, file)
    dst = os.path.join(train_dir, file)
    shutil.move(src, dst)

for file in os.listdir(meningioma_train_path):
    src = os.path.join(meningioma_train_path, file)
    dst = os.path.join(train_dir, file)
    shutil.move(src, dst)

for file in os.listdir(notumor_train_path):
    src = os.path.join(notumor_train_path, file)
    dst = os.path.join(train_dir, file)
    shutil.move(src, dst)

for file in os.listdir(pituitary_train_path):
    src = os.path.join(pituitary_train_path, file)
    dst = os.path.join(train_dir, file)
    shutil.move(src, dst)

print("Train splitting completed successfully!")

# Define Test Directory

import os
import shutil

# Define the train directory
test_dir = '/content/drive/MyDrive/BrainTumorClassification/Dataset/Test'

# Ensure the existence of the train and test directories
os.makedirs(test_dir, exist_ok=True)

# Move the remaining files to the train directory
for file in os.listdir(glioma_test_path):
    src = os.path.join(glioma_test_path, file)
    dst = os.path.join(test_dir, file)
    shutil.move(src, dst)

for file in os.listdir(meningioma_test_path):
    src = os.path.join(meningioma_test_path, file)
    dst = os.path.join(test_dir, file)
    shutil.move(src, dst)

for file in os.listdir(notumor_test_path):
    src = os.path.join(notumor_test_path, file)
    dst = os.path.join(test_dir, file)
    shutil.move(src, dst)

for file in os.listdir(pituitary_test_path):
    src = os.path.join(pituitary_test_path, file)
    dst = os.path.join(test_dir, file)
    shutil.move(src, dst)

print("Train splitting completed successfully!")
