==> 2_Setup_Image_Paths_and_labels.ipynb <==
# Mount Google Drive

from google.colab import drive
drive.mount('/content/drive')

# Setup Image Paths and Labels

## For Train Set

import numpy as np
import pandas as pd
import os

Id=[]
# Append directory name and filenames
for dirname, _, filenames in os.walk('/content/drive/MyDrive/BrainTumorClassification/Dataset/Train'):
    for filename in filenames:
        Id.append(os.path.join(dirname, filename))
Id[:10]

# Convert filenames into dataframe
train = pd.DataFrame()
train = train.assign(filename=Id)
train.head()

# Add new column 'label' and process labels for images
train['label'] = train['filename']

for i in range(len(train)):
    if 'glioma' in train['filename'][i]:
        train['label'][i] = 'glioma'
    if 'meningioma' in train['filename'][i]:
        train['label'][i] = 'meningioma'
    if 'notumor' in train['filename'][i]:
        train['label'][i] = 'notumor'
    if 'pituitary' in train['filename'][i]:
        train['label'][i] = 'pituitary'

train.head()

train.tail()

train['label'].value_counts()

## For Test Set

Id = []
for dirname, _, filenames in os.walk('/content/drive/MyDrive/BrainTumorClassification/Dataset/Test'):
    for filename in filenames:
        Id.append(os.path.join(dirname, filename))
Id[:10]

# Convert image id's to pandas dataframe
test = pd.DataFrame().assign(filename=Id)
test.head()

test.tail()

# Add new column 'label' and process labels for images
test['label'] = test['filename']

for i in range(len(test)):
    if 'glioma' in test['filename'][i]:
        test['label'][i] = 'glioma'
    if 'meningioma' in test['filename'][i]:
        test['label'][i] = 'meningioma'
    if 'notumor' in test['filename'][i]:
        test['label'][i] = 'notumor'
    if 'pituitary' in test['filename'][i]:
        test['label'][i] = 'pituitary'

test.head()

test.tail()

test['label'].value_counts()

# Save Image paths and labels as CSV

# Save train and test dataframe as csv for future use
train.to_csv('train.csv', index=False)
test.to_csv('test.csv', index=False)

from google.colab import drive
from google.colab import files
import shutil

# Specify the destination directory
destination_folder = '/content/drive/MyDrive/BrainTumorClassification'
# Upload train.csv
shutil.move('train.csv', destination_folder)
# Upload test.csv
shutil.move('test.csv', destination_folder)

print("Files uploaded successfully!")
