==> 3_Train_ResNet_18_Model.ipynb <==
# Mount Google Drive

from google.colab import drive
drive.mount('/content/drive')

# Process and Transform Images

import pandas as pd
from torch.utils.data import Dataset
from torchvision import transforms
from PIL import Image

class BrainTumorDataset(Dataset):
    def __init__(self, csv_file, transform=None):
        self.data = pd.read_csv(csv_file)
        self.transform = transform

    def __len__(self):
        return len(self.data)

    def __getitem__(self, idx):
        img_path = self.data.iloc[idx, 0]
        label = self.data.iloc[idx, 1]

        # Encode labels
        label_dict = {'notumor': 0, 'glioma': 1, 'meningioma': 2, 'pituitary': 3}
        encoded_label = label_dict[label]

        # Load image
        image = Image.open(img_path).convert('RGB')

        if self.transform:
            image = self.transform(image)

        return image, encoded_label

# Define data transformation
data_transform = transforms.Compose([
    transforms.Resize((224, 224)),  # Resize images to the input size expected by the CNN model
    transforms.ToTensor(),
    transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]), # Normalize pixel values
])

# get your customized dataset
train_set = BrainTumorDataset(csv_file='/content/drive/MyDrive/BrainTumorClassification/train.csv', transform=data_transform)
test_set = BrainTumorDataset(csv_file='/content/drive/MyDrive/BrainTumorClassification/test.csv', transform=data_transform)

# Print set sizes
print(len(train_set))
print(len(test_set))

from torch.utils.data import Dataset, DataLoader

batch_size=16

# Create DataLoader
train_loader = DataLoader(train_set, batch_size=batch_size, shuffle=True)
test_loader = DataLoader(test_set, batch_size=batch_size, shuffle=False)

# Print loader
print(len(train_loader))
print(len(test_loader))

# Train ResNet-18 Model

import torch
import torchvision
import torch.nn as nn
import torch.optim as optim

# Define the model
model = torchvision.models.resnet18(pretrained=True)

# Modify the last fully connected layer to fit our 4 classes
num_ftrs = model.fc.in_features
model.fc = nn.Linear(num_ftrs, 4)
model

# Define device
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model.to(device)

# Check if the model is on GPU
if next(model.parameters()).is_cuda:
    print("Model is currently on GPU")
else:
    print("Model is currently on CPU")

# Initialize the loss function
loss_fn = nn.CrossEntropyLoss()
# Adam Optimizer
optimizer = optim.Adam(model.parameters(), lr=0.001)

train_losses = []
train_accuracies = []
test_losses = []
test_accuracies = []

# Train the model
num_epochs = 10
for epoch in range(num_epochs):
    model.train()  # Set the model to training mode
    running_train_loss = 0.0
    correct_train = 0
    total_train = 0

    # Training phase
    for images, labels in train_loader:
        images, labels = images.to(device), labels.to(device)
        # Zero the parameter gradients
        optimizer.zero_grad()
        # Forward pass
        outputs = model(images)
        loss = loss_fn(outputs, labels)
        # Backward pass and optimize
        loss.backward()
        optimizer.step()

        # Compute statistics
        running_train_loss += loss.item()
        _, predicted_train = torch.max(outputs.data, 1)
        total_train += labels.size(0)
        correct_train += (predicted_train == labels).sum().item()

    # Testing phase
    model.eval()  # Set the model to evaluation mode
    running_test_loss = 0.0
    correct_test = 0
    total_test = 0

    with torch.no_grad():
        for images, labels in test_loader:
            images, labels = images.to(device), labels.to(device)

            # Forward pass
            outputs = model(images)
            loss = loss_fn(outputs, labels)

            # Compute statistics
            running_test_loss += loss.item()
            _, predicted_test = torch.max(outputs.data, 1)
            total_test += labels.size(0)
            correct_test += (predicted_test == labels).sum().item()

    # Store results in global lists
    train_losses.append(running_train_loss / len(train_loader))
    train_accuracies.append(100 * correct_train / total_train)
    test_losses.append(running_test_loss / len(test_loader))
    test_accuracies.append(100 * correct_test / total_test)

    # Print statistics
    print(f"Epoch [{epoch+1}/{num_epochs}], "
          f"Train Loss: {train_losses[-1]:.4f}, "
          f"Train Accuracy: {train_accuracies[-1]:.2f}%, "
          f"Test Loss: {test_losses[-1]:.4f}, "
          f"Test Accuracy: {test_accuracies[-1]:.2f}%")
